Restructure
This commit restructures the tool into discrete modules
